+++ src/main.js
@@ -6,6 +6,7 @@ import { version } from '../package.json'
 import * as Sentry from '@sentry/browser'
 import * as Integrations from '@sentry/integrations'
 import VuePageTransition from 'vue-page-transition'
+import VueModal from 'vue-js-modal'
 
 import '@/sass/globals.scss'
 import '@/sass/fonts.scss'
@@ -13,9 +14,10 @@ import '@/sass/fonts.scss'
 Vue.config.productionTip = false
 
 Vue.use(VuePageTransition)
+Vue.use(VueModal, { dynamic: true, injectModalsContainer: true })
 
 Sentry.init({
-  dsn: 'https://ce5b907a4c294be1be5793b281daa9cc@sentry.io/1869953',
+  dsn: process.env.SENTRY_DSN,
   maxBreadcrumbs: 50,
   debug: process.env.NODE_ENV === 'production' ? false : true,
   release: 'sparkboard@' + version,