+++ src/store/index.js
@@ -1,11 +1,19 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
+import board from '@/store/modules/board'
 
 Vue.use(Vuex)
 
+const debug = process.env.NODE_ENV !== 'production'
+
 export default new Vuex.Store({
-  state: {},
+  state: {
+    boards: [],
+  },
   mutations: {},
   actions: {},
-  modules: {},
+  modules: {
+    board,
+  },
+  strict: debug,
 })